@@ -1,232 +1,292 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
 
-	<groupId>de.viadee.camunda</groupId>
-	<artifactId>camunda-kafka-polling-client-parent</artifactId>
-	<version>1.0.2</version>
-	<packaging>pom</packaging>
+    <groupId>de.viadee.camunda</groupId>
+    <artifactId>camunda-kafka-polling-client-parent</artifactId>
+    <version>1.1.0</version>
+    <packaging>pom</packaging>
 
-	<modules>
-		<module>camunda-kafka-model</module>
-		<module>camunda-kafka-polling-client</module>
-	</modules>
+    <modules>
+        <module>camunda-kafka-model</module>
+        <module>camunda-kafka-polling-client</module>
+    </modules>
 
-	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
 
-		<spring-boot.version>2.0.0.RELEASE</spring-boot.version>
-		<camunda.version>7.8.0</camunda.version>
-		<kafka.version>1.0.0</kafka.version>
-		<jackson.version>2.9.8</jackson.version>
-		<postgresql.version>42.2.1</postgresql.version>
-		<h2.version>1.4.196</h2.version>
-	</properties>
+        <spring-boot.version>2.1.3.RELEASE</spring-boot.version>
+        <camunda.version>7.10.0</camunda.version>
+        <kafka.version>1.0.0</kafka.version>
+        <jackson.version>2.9.8</jackson.version>
+        <postgresql.version>42.2.5</postgresql.version>
+        <junit.version>5.4.0</junit.version>
+        <mockito.version>2.25.0</mockito.version>
+        <commons-collections4.version>4.3</commons-collections4.version>
+        <commons-io.version>2.6</commons-io.version>
+    </properties>
 
-	<url>https://github.com/viadee/camunda-kafka-polling-client</url>
-	<organization>
-		<name>viadee Unternehmensberatung AG</name>
-		<url>www.viadee.de</url>
-	</organization>
+    <url>https://github.com/viadee/camunda-kafka-polling-client</url>
+    <organization>
+        <name>viadee Unternehmensberatung AG</name>
+        <url>www.viadee.de</url>
+    </organization>
 
-	<name>camunda-kafka-polling-client-parent</name>
-	<description>
-		A small Java-based toolset to gather event data from a Camunda BPM engine and store it in Apache Kafka.
-		Repository contains the camunda-kafka-polling-client and the camunda-kafka-model, to make reuse of the data
-		structures easy.
-		This tooling is meant to be used as a part of the viadee Process Warehouse or the bpmn.ai concepts.
-	</description>
-	<inceptionYear>2018</inceptionYear>
+    <name>camunda-kafka-polling-client-parent</name>
+    <description>
+        A small Java-based toolset to gather event data from a Camunda BPM engine and store it in Apache Kafka.
+        Repository contains the camunda-kafka-polling-client and the camunda-kafka-model, to make reuse of the data
+        structures easy.
+        This tooling is meant to be used as a part of the viadee Process Warehouse or the bpmn.ai concepts.
+    </description>
+    <inceptionYear>2018</inceptionYear>
 
-	<licenses>
-		<license>
-			<name>BSD-3 Clause</name>
-			<url>https://opensource.org/licenses/BSD-3-Clause</url>
-		</license>
-	</licenses>
+    <licenses>
+        <license>
+            <name>BSD-3 Clause</name>
+            <url>https://opensource.org/licenses/BSD-3-Clause</url>
+        </license>
+    </licenses>
 
-	<ciManagement>
-		<url>https://travis-ci.org/viadee/camunda-kafka-polling-client</url>
-		<system>Travis</system>
-	</ciManagement>
+    <ciManagement>
+        <url>https://travis-ci.org/viadee/camunda-kafka-polling-client</url>
+        <system>Travis</system>
+    </ciManagement>
 
-	<issueManagement>
-		<url>https://github.com/viadee/camunda-kafka-polling-client/issues</url>
-		<system>GitHub</system>
-	</issueManagement>
+    <issueManagement>
+        <url>https://github.com/viadee/camunda-kafka-polling-client/issues</url>
+        <system>GitHub</system>
+    </issueManagement>
 
-	<scm>
-		<connection>https://github.com/viadee/camunda-kafka-polling-client</connection>
-		<url>GitHub</url>
-	</scm>
+    <scm>
+        <connection>https://github.com/viadee/camunda-kafka-polling-client</connection>
+        <url>GitHub</url>
+    </scm>
 
-	<developers>
-		<developer>
-			<name>Michael Twelkemeier</name>
-			<email>Michael.Twelkemeier@viadee.de</email>
-			<organization>viadee Unternehmensberatung AG</organization>
-			<organizationUrl>http://www.viadee.de</organizationUrl>
-		</developer>
-		<developer>
-			<name>Jan-Philipp Friedenstab</name>
-			<email>Jan-Philipp.Friedenstab@viadee.de</email>
-			<organization>viadee Unternehmensberatung AG</organization>
-			<organizationUrl>http://www.viadee.de</organizationUrl>
-		</developer>
-		<developer>
-			<name>Sebastian Sirch</name>
-			<email>Sebastian.Sirch@viadee.de</email>
-			<organization>viadee Unternehmensberatung AG</organization>
-			<organizationUrl>http://www.viadee.de</organizationUrl>
-		</developer>
-		<developer>
-			<name>Mario Micudaj</name>
-			<email>Mario.Micudaj@viadee.de</email>
-			<organization>viadee Unternehmensberatung AG</organization>
-			<organizationUrl>http://www.viadee.de</organizationUrl>
-		</developer>
-		<developer>
-			<name>Thorben Hellweg</name>
-			<email>Thorben.Hellweg@viadee.de</email>
-			<organization>viadee Unternehmensberatung AG</organization>
-			<organizationUrl>http://www.viadee.de</organizationUrl>
-		</developer>
-		<developer>
-			<name>Caroline Methner</name>
-			<email>Caroline.Methner@viadee.de</email>
-			<organization>viadee Unternehmensberatung AG</organization>
-			<organizationUrl>http://www.viadee.de</organizationUrl>
-		</developer>
-		<developer>
-			<name>Frank Koehne</name>
-			<email>Frank.Koehne@viadee.de</email>
-			<organization>viadee Unternehmensberatung AG</organization>
-			<organizationUrl>http://www.viadee.de</organizationUrl>
-		</developer>
-	</developers>
+    <developers>
+        <developer>
+            <name>Michael Twelkemeier</name>
+            <email>Michael.Twelkemeier@viadee.de</email>
+            <organization>viadee Unternehmensberatung AG</organization>
+            <organizationUrl>http://www.viadee.de</organizationUrl>
+        </developer>
+        <developer>
+            <name>Jan-Philipp Friedenstab</name>
+            <email>Jan-Philipp.Friedenstab@viadee.de</email>
+            <organization>viadee Unternehmensberatung AG</organization>
+            <organizationUrl>http://www.viadee.de</organizationUrl>
+        </developer>
+        <developer>
+            <name>Sebastian Sirch</name>
+            <email>Sebastian.Sirch@viadee.de</email>
+            <organization>viadee Unternehmensberatung AG</organization>
+            <organizationUrl>http://www.viadee.de</organizationUrl>
+        </developer>
+        <developer>
+            <name>Mario Micudaj</name>
+            <email>Mario.Micudaj@viadee.de</email>
+            <organization>viadee Unternehmensberatung AG</organization>
+            <organizationUrl>http://www.viadee.de</organizationUrl>
+        </developer>
+        <developer>
+            <name>Thorben Hellweg</name>
+            <email>Thorben.Hellweg@viadee.de</email>
+            <organization>viadee Unternehmensberatung AG</organization>
+            <organizationUrl>http://www.viadee.de</organizationUrl>
+        </developer>
+        <developer>
+            <name>Caroline Methner</name>
+            <email>Caroline.Methner@viadee.de</email>
+            <organization>viadee Unternehmensberatung AG</organization>
+            <organizationUrl>http://www.viadee.de</organizationUrl>
+        </developer>
+        <developer>
+            <name>Frank Koehne</name>
+            <email>Frank.Koehne@viadee.de</email>
+            <organization>viadee Unternehmensberatung AG</organization>
+            <organizationUrl>http://www.viadee.de</organizationUrl>
+        </developer>
+    </developers>
 
-	<profiles>
-		<profile>
-			<id>release-sign-artifacts</id>
-			<activation>
-				<property>
-					<name>performRelease</name>
-					<value>true</value>
-				</property>
-			</activation>
-			<build>
-				<plugins>
-					<plugin>
-						<groupId>org.apache.maven.plugins</groupId>
-						<artifactId>maven-gpg-plugin</artifactId>
-						<version>1.6</version>
-						<executions>
-							<execution>
-								<id>sign-artifacts</id>
-								<phase>verify</phase>
-								<goals>
-									<goal>sign</goal>
-								</goals>
-							</execution>
-						</executions>
-					</plugin>
-				</plugins>
-			</build>
-		</profile>
-	</profiles>
+    <profiles>
+        <profile>
+            <id>release-sign-artifacts</id>
+            <activation>
+                <property>
+                    <name>performRelease</name>
+                    <value>true</value>
+                </property>
+            </activation>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-gpg-plugin</artifactId>
+                        <version>1.6</version>
+                        <executions>
+                            <execution>
+                                <id>sign-artifacts</id>
+                                <phase>verify</phase>
+                                <goals>
+                                    <goal>sign</goal>
+                                </goals>
+                            </execution>
+                        </executions>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+    </profiles>
 
-	<distributionManagement>
-		<site>
-			<id>website</id>
-			<url>file://${basedir}/docs/MavenSite</url>
-		</site>
-		<snapshotRepository>
-			<id>ossrh</id>
-			<url>https://oss.sonatype.org/content/repositories/snapshots</url>
-		</snapshotRepository>
-		<repository>
-			<id>ossrh</id>
-			<url>https://oss.sonatype.org/service/local/staging/deploy/maven2</url>
-		</repository>
-	</distributionManagement>
+    <distributionManagement>
+        <site>
+            <id>website</id>
+            <url>file://${basedir}/docs/MavenSite</url>
+        </site>
+        <snapshotRepository>
+            <id>ossrh</id>
+            <url>https://oss.sonatype.org/content/repositories/snapshots</url>
+        </snapshotRepository>
+        <repository>
+            <id>ossrh</id>
+            <url>https://oss.sonatype.org/service/local/staging/deploy/maven2</url>
+        </repository>
+    </distributionManagement>
 
-	<dependencyManagement>
-		<dependencies>
-			<dependency>
-				<groupId>org.springframework.boot</groupId>
-				<artifactId>spring-boot-dependencies</artifactId>
-				<version>${spring-boot.version}</version>
-				<type>pom</type>
-				<scope>import</scope>
-			</dependency>
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-dependencies</artifactId>
+                <version>${spring-boot.version}</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
 
-			<dependency>
-				<groupId>org.camunda.bpm</groupId>
-				<artifactId>camunda-bom</artifactId>
-				<version>${camunda.version}</version>
-				<scope>import</scope>
-				<type>pom</type>
-			</dependency>
+            <dependency>
+                <groupId>org.camunda.bpm</groupId>
+                <artifactId>camunda-bom</artifactId>
+                <version>${camunda.version}</version>
+                <scope>import</scope>
+                <type>pom</type>
+            </dependency>
 
-			<dependency>
-				<groupId>com.fasterxml.jackson.core</groupId>
-				<artifactId>jackson-databind</artifactId>
-				<version>${jackson.version}</version>
-			</dependency>
-			<dependency>
-				<groupId>com.fasterxml.jackson.core</groupId>
-				<artifactId>jackson-core</artifactId>
-				<version>${jackson.version}</version>
-			</dependency>
-			<dependency>
-				<groupId>com.fasterxml.jackson.core</groupId>
-				<artifactId>jackson-annotations</artifactId>
-				<version>${jackson.version}</version>
-			</dependency>
+            <dependency>
+                <groupId>com.fasterxml.jackson.core</groupId>
+                <artifactId>jackson-databind</artifactId>
+                <version>${jackson.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.fasterxml.jackson.core</groupId>
+                <artifactId>jackson-core</artifactId>
+                <version>${jackson.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.fasterxml.jackson.core</groupId>
+                <artifactId>jackson-annotations</artifactId>
+                <version>${jackson.version}</version>
+            </dependency>
 
-			<dependency>
-				<groupId>org.apache.kafka</groupId>
-				<artifactId>kafka-clients</artifactId>
-				<version>${kafka.version}</version>
-			</dependency>
+            <dependency>
+                <groupId>org.apache.kafka</groupId>
+                <artifactId>kafka-clients</artifactId>
+                <version>${kafka.version}</version>
+            </dependency>
 
-			<dependency>
-				<groupId>org.postgresql</groupId>
-				<artifactId>postgresql</artifactId>
-				<version>${postgresql.version}</version>
-			</dependency>
-			<dependency>
-				<groupId>com.h2database</groupId>
-				<artifactId>h2</artifactId>
-				<version>${h2.version}</version>
-			</dependency>
-		</dependencies>
-	</dependencyManagement>
+            <dependency>
+                <groupId>org.postgresql</groupId>
+                <artifactId>postgresql</artifactId>
+                <version>${postgresql.version}</version>
+            </dependency>
 
-	<build>
-		<finalName>${project.artifactId}</finalName>
-		<pluginManagement>
-			<plugins>
-				<plugin>
-					<groupId>org.springframework.boot</groupId>
-					<artifactId>spring-boot-maven-plugin</artifactId>
-					<version>${spring-boot.version}</version>
-					<executions>
-						<execution>
-							<goals>
-								<goal>repackage</goal>
-							</goals>
-						</execution>
-					</executions>
-				</plugin>
+            <dependency>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-starter-test</artifactId>
+                <version>${spring-boot.version}</version>
+                <exclusions>
+                    <!-- Since using JUnit 5, exclude JUnit 4 dependencies to ensure only one is used -->
+                    <exclusion>
+                        <groupId>junit</groupId>
+                        <artifactId>junit</artifactId>
+                    </exclusion>
+                    <exclusion>
+                        <groupId>org.mockito</groupId>
+                        <artifactId>mockito-core</artifactId>
+                    </exclusion>
+                </exclusions>
+            </dependency>
 
-			</plugins>
-		</pluginManagement>
-	</build>
+            <dependency>
+                <groupId>org.junit.jupiter</groupId>
+                <artifactId>junit-jupiter</artifactId>
+                <version>${junit.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.mockito</groupId>
+                <artifactId>mockito-core</artifactId>
+                <version>${mockito.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.mockito</groupId>
+                <artifactId>mockito-junit-jupiter</artifactId>
+                <version>${mockito.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.apache.commons</groupId>
+                <artifactId>commons-collections4</artifactId>
+                <version>${commons-collections4.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>commons-io</groupId>
+                <artifactId>commons-io</artifactId>
+                <version>${commons-io.version}</version>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
+
+    <build>
+        <finalName>${project.artifactId}</finalName>
+
+        <pluginManagement>
+            <plugins>
+                <plugin>
+                    <groupId>net.revelc.code.formatter</groupId>
+                    <artifactId>formatter-maven-plugin</artifactId>
+                    <version>2.8.1</version>
+                    <configuration>
+                        <configFile>${project.basedir}/code-formatter.xml</configFile>
+                    </configuration>
+                    <executions>
+                        <execution>
+                            <goals>
+                                <goal>validate</goal>
+                            </goals>
+                            <phase>verify</phase>
+                        </execution>
+                    </executions>
+                </plugin>
+
+                <plugin>
+                    <groupId>org.springframework.boot</groupId>
+                    <artifactId>spring-boot-maven-plugin</artifactId>
+                    <version>${spring-boot.version}</version>
+                    <executions>
+                        <execution>
+                            <goals>
+                                <goal>repackage</goal>
+                            </goals>
+                        </execution>
+                    </executions>
+                </plugin>
+            </plugins>
+        </pluginManagement>
+    </build>
 </project>
