@@ -4,12 +4,18 @@ package de.viadee.camunda.kafka.event;
 import java.util.Date;
 
 /**
- * <p>ScopeInstanceEvent class.</p>
- *
- * @author viadee
+ * <p>
+ * ScopeInstanceEvent class.
+ * </p>
+ *
+ * 
+ * 
+ * @author viadee
+ * 
  * @version $Id: $Id
  */
 public class ScopeInstanceEvent extends HistoryEvent {
+
     private Long durationInMillis;
     private Date startTime;
     private Date endTime;
