@@ -1,10 +1,5 @@
 package de.viadee.camunda.kafka.pollingclient.config;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
 import de.viadee.camunda.kafka.pollingclient.config.properties.ApplicationProperties;
 import de.viadee.camunda.kafka.pollingclient.job.runtime.RuntimeDataPollingJob;
 import de.viadee.camunda.kafka.pollingclient.job.runtime.RuntimeDataPollingService;
@@ -12,9 +7,15 @@ import de.viadee.camunda.kafka.pollingclient.service.event.EventService;
 import de.viadee.camunda.kafka.pollingclient.service.lastpolled.LastPolledService;
 import de.viadee.camunda.kafka.pollingclient.service.lastpolled.filebased.FilebasedLastPolledServiceImpl;
 import de.viadee.camunda.kafka.pollingclient.service.polling.PollingService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 
 /**
- * <p>RuntimeDataPollingConfiguration class.</p>
+ * <p>
+ * RuntimeDataPollingConfiguration class.
+ * </p>
  *
  * @author viadee
  * @version $Id: $Id
@@ -32,7 +33,9 @@ public class RuntimeDataPollingConfiguration {
     private EventService eventService;
 
     /**
-     * <p>runtimeDataLastPolledService.</p>
+     * <p>
+     * runtimeDataLastPolledService.
+     * </p>
      *
      * @return a {@link de.viadee.camunda.kafka.pollingclient.service.lastpolled.LastPolledService} object.
      */
@@ -42,7 +45,9 @@ public class RuntimeDataPollingConfiguration {
     }
 
     /**
-     * <p>runtimeDataPollingService.</p>
+     * <p>
+     * runtimeDataPollingService.
+     * </p>
      *
      * @return a {@link de.viadee.camunda.kafka.pollingclient.job.runtime.RuntimeDataPollingService} object.
      */
@@ -52,7 +57,9 @@ public class RuntimeDataPollingConfiguration {
     }
 
     /**
-     * <p>runtimeDataPollingJob.</p>
+     * <p>
+     * runtimeDataPollingJob.
+     * </p>
      *
      * @return a {@link de.viadee.camunda.kafka.pollingclient.job.runtime.RuntimeDataPollingJob} object.
      */
