@@ -1,13 +1,14 @@
 package de.viadee.camunda.kafka.pollingclient.config.properties;
 
+import org.springframework.format.annotation.DateTimeFormat;
+
 import java.io.File;
 import java.util.Date;
 
-
-import org.springframework.format.annotation.DateTimeFormat;
-
 /**
- * <p>PollingProperties class.</p>
+ * <p>
+ * PollingProperties class.
+ * </p>
  *
  * @author viadee
  * @version $Id: $Id
@@ -16,7 +17,8 @@ import org.springframework.format.annotation.DateTimeFormat;
 public class PollingProperties {
 
     /**
-     * Initial timestamp to start polling with in case no polling has been performed before. (Default: Start timestamp of polling client)
+     * Initial timestamp to start polling with in case no polling has been performed before. (Default: Start timestamp
+     * of polling client)
      */
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private Date initialTimestamp = new Date();
@@ -35,9 +37,10 @@ public class PollingProperties {
      * Enable/disable polling
      */
     private boolean enabled;
-    
+
     /**
-     * Initial timestamp to start polling with in case no polling has been performed before. (Default: Start timestamp of polling client)
+     * Initial timestamp to start polling with in case no polling has been performed before. (Default: Start timestamp
+     * of polling client)
      */
     @java.lang.SuppressWarnings("all")
     public Date getInitialTimestamp() {
@@ -69,7 +72,8 @@ public class PollingProperties {
     }
 
     /**
-     * Initial timestamp to start polling with in case no polling has been performed before. (Default: Start timestamp of polling client)
+     * Initial timestamp to start polling with in case no polling has been performed before. (Default: Start timestamp
+     * of polling client)
      */
     @java.lang.SuppressWarnings("all")
     public void setInitialTimestamp(final Date initialTimestamp) {
