@@ -3,7 +3,9 @@ package de.viadee.camunda.kafka.pollingclient.job.repository;
 import org.springframework.scheduling.annotation.Scheduled;
 
 /**
- * <p>RepositoryDataPollingJob class.</p>
+ * <p>
+ * RepositoryDataPollingJob class.
+ * </p>
  *
  * @author viadee
  * @version $Id: $Id
@@ -13,16 +15,21 @@ public class RepositoryDataPollingJob {
     private final RepositoryDataPollingService repositoryDataPollingService;
 
     /**
-     * <p>Constructor for RepositoryDataPollingJob.</p>
+     * <p>
+     * Constructor for RepositoryDataPollingJob.
+     * </p>
      *
-     * @param repositoryDataPollingService a {@link de.viadee.camunda.kafka.pollingclient.job.repository.RepositoryDataPollingService} object.
+     * @param repositoryDataPollingService
+     *            a {@link de.viadee.camunda.kafka.pollingclient.job.repository.RepositoryDataPollingService} object.
      */
     public RepositoryDataPollingJob(final RepositoryDataPollingService repositoryDataPollingService) {
         this.repositoryDataPollingService = repositoryDataPollingService;
     }
 
     /**
-     * <p>executeScheduled.</p>
+     * <p>
+     * executeScheduled.
+     * </p>
      */
     @Scheduled(initialDelay = 500L, fixedDelayString = "${polling.repository-data.interval-in-ms}")
     public void executeScheduled() {
