@@ -1,21 +1,22 @@
 package de.viadee.camunda.kafka.pollingclient.service.event.kafka;
 
-import static org.apache.commons.lang3.StringUtils.*;
-
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import de.viadee.camunda.kafka.event.DeploymentEvent;
+import de.viadee.camunda.kafka.event.HistoryEvent;
+import de.viadee.camunda.kafka.pollingclient.config.properties.ApplicationProperties;
+import de.viadee.camunda.kafka.pollingclient.service.event.EventService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import de.viadee.camunda.kafka.pollingclient.config.properties.ApplicationProperties;
-import de.viadee.camunda.kafka.pollingclient.service.event.EventService;
-import de.viadee.camunda.kafka.event.DeploymentEvent;
-import de.viadee.camunda.kafka.event.HistoryEvent;
+import static org.apache.commons.lang3.StringUtils.removeEnd;
+import static org.apache.commons.lang3.StringUtils.uncapitalize;
 
 /**
- * <p>KafkaEventServiceImpl class.</p>
+ * <p>
+ * KafkaEventServiceImpl class.
+ * </p>
  *
  * @author viadee
  * @version $Id: $Id
@@ -31,10 +32,14 @@ public class KafkaEventServiceImpl implements EventService {
 
     @Autowired
     /**
-     * <p>Constructor for KafkaEventServiceImpl.</p>
+     * <p>
+     * Constructor for KafkaEventServiceImpl.
+     * </p>
      *
-     * @param kafkaTemplate a {@link org.springframework.kafka.core.KafkaTemplate} object.
-     * @param properties a {@link de.viadee.camunda.kafka.pollingclient.config.properties.ApplicationProperties} object.
+     * @param kafkaTemplate
+     *            a {@link org.springframework.kafka.core.KafkaTemplate} object.
+     * @param properties
+     *            a {@link de.viadee.camunda.kafka.pollingclient.config.properties.ApplicationProperties} object.
      */
     public KafkaEventServiceImpl(KafkaTemplate<String, String> kafkaTemplate, ApplicationProperties properties) {
         this.kafkaTemplate = kafkaTemplate;
