@@ -1,14 +1,16 @@
 // Generated by delombok at Fri Jan 04 11:18:40 CET 2019
 package de.viadee.camunda.kafka.pollingclient.service.polling.rest.response;
 
-import java.util.Date;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import java.util.Date;
 
 /**
  * Response structure of Camunda REST API <code>GET /deployment</code>
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class GetDeploymentResponse {
+
     private String id;
     private String name;
     private String source;
