@@ -1,14 +1,16 @@
 // Generated by delombok at Fri Jan 04 11:18:40 CET 2019
 package de.viadee.camunda.kafka.pollingclient.service.polling.rest.response;
 
-import java.util.Date;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import java.util.Date;
 
 /**
  * Response structure of Camunda REST API <code>GET /history/activity-instance</code>
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class GetHistoricActivityInstanceRespone {
+
     private String id;
     private String parentActivityInstanceId;
     private String activityId;
@@ -222,6 +224,15 @@ public class GetHistoricActivityInstanceRespone {
     @java.lang.Override
     @java.lang.SuppressWarnings("all")
     public java.lang.String toString() {
-        return "GetHistoricActivityInstanceRespone(id=" + this.getId() + ", parentActivityInstanceId=" + this.getParentActivityInstanceId() + ", activityId=" + this.getActivityId() + ", activityName=" + this.getActivityName() + ", activityType=" + this.getActivityType() + ", processDefinitionKey=" + this.getProcessDefinitionKey() + ", processDefinitionId=" + this.getProcessDefinitionId() + ", processInstanceId=" + this.getProcessInstanceId() + ", executionId=" + this.getExecutionId() + ", taskId=" + this.getTaskId() + ", assignee=" + this.getAssignee() + ", calledProcessInstanceId=" + this.getCalledProcessInstanceId() + ", calledCaseInstanceId=" + this.getCalledCaseInstanceId() + ", startTime=" + this.getStartTime() + ", endTime=" + this.getEndTime() + ", durationInMillis=" + this.getDurationInMillis() + ", canceled=" + this.getCanceled() + ", completeScope=" + this.getCompleteScope() + ", tenantId=" + this.getTenantId() + ")";
+        return "GetHistoricActivityInstanceRespone(id=" + this.getId() + ", parentActivityInstanceId="
+                + this.getParentActivityInstanceId() + ", activityId=" + this.getActivityId() + ", activityName="
+                + this.getActivityName() + ", activityType=" + this.getActivityType() + ", processDefinitionKey="
+                + this.getProcessDefinitionKey() + ", processDefinitionId=" + this.getProcessDefinitionId()
+                + ", processInstanceId=" + this.getProcessInstanceId() + ", executionId=" + this.getExecutionId()
+                + ", taskId=" + this.getTaskId() + ", assignee=" + this.getAssignee() + ", calledProcessInstanceId="
+                + this.getCalledProcessInstanceId() + ", calledCaseInstanceId=" + this.getCalledCaseInstanceId()
+                + ", startTime=" + this.getStartTime() + ", endTime=" + this.getEndTime() + ", durationInMillis="
+                + this.getDurationInMillis() + ", canceled=" + this.getCanceled() + ", completeScope="
+                + this.getCompleteScope() + ", tenantId=" + this.getTenantId() + ")";
     }
 }
