@@ -1,17 +1,19 @@
 // Generated by delombok at Fri Jan 04 11:18:40 CET 2019
 package de.viadee.camunda.kafka.pollingclient.service.polling.rest.response;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import org.apache.commons.lang3.StringUtils;
+
 import java.util.Date;
 import java.util.Map;
 import java.util.Objects;
-import org.apache.commons.lang3.StringUtils;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 /**
  * Response structure of Camunda REST API <code>GET /history/detail for type=variableUpdate</code>
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class GetHistoricDetailVariableUpdateResponse {
+
     private String id;
     private String type;
     private String processDefinitionKey;
@@ -38,7 +40,8 @@ public class GetHistoricDetailVariableUpdateResponse {
      * If {@link #valueInfo} is available as Map-Data, retrieve value of given key as String.
      */
     public String getValueInfoEntry(String key) {
-        return valueInfo instanceof Map ? StringUtils.trimToNull(Objects.toString(((Map) valueInfo).get(key), null)) : null;
+        return valueInfo instanceof Map ? StringUtils.trimToNull(Objects.toString(((Map) valueInfo).get(key), null))
+                : null;
     }
 
     @java.lang.SuppressWarnings("all")
