@@ -1,14 +1,16 @@
 // Generated by delombok at Fri Jan 04 11:18:40 CET 2019
 package de.viadee.camunda.kafka.pollingclient.service.polling.rest.response;
 
-import java.util.Date;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import java.util.Date;
 
 /**
  * Response structure of Camunda REST API <code>GET /history/process-instance</code>
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class GetHistoricProcessInstanceResponse {
+
     private String id;
     private String superProcessInstanceId;
     private String superCaseInstanceId;
@@ -200,6 +202,15 @@ public class GetHistoricProcessInstanceResponse {
     @java.lang.Override
     @java.lang.SuppressWarnings("all")
     public java.lang.String toString() {
-        return "GetHistoricProcessInstanceResponse(id=" + this.getId() + ", superProcessInstanceId=" + this.getSuperProcessInstanceId() + ", superCaseInstanceId=" + this.getSuperCaseInstanceId() + ", caseInstanceId=" + this.getCaseInstanceId() + ", processDefinitionName=" + this.getProcessDefinitionName() + ", processDefinitionKey=" + this.getProcessDefinitionKey() + ", processDefinitionVersion=" + this.getProcessDefinitionVersion() + ", processDefinitionId=" + this.getProcessDefinitionId() + ", businessKey=" + this.getBusinessKey() + ", startTime=" + this.getStartTime() + ", endTime=" + this.getEndTime() + ", durationInMillis=" + this.getDurationInMillis() + ", startUserId=" + this.getStartUserId() + ", startActivityId=" + this.getStartActivityId() + ", deleteReason=" + this.getDeleteReason() + ", tenantId=" + this.getTenantId() + ", state=" + this.getState() + ")";
+        return "GetHistoricProcessInstanceResponse(id=" + this.getId() + ", superProcessInstanceId="
+                + this.getSuperProcessInstanceId() + ", superCaseInstanceId=" + this.getSuperCaseInstanceId()
+                + ", caseInstanceId=" + this.getCaseInstanceId() + ", processDefinitionName="
+                + this.getProcessDefinitionName() + ", processDefinitionKey=" + this.getProcessDefinitionKey()
+                + ", processDefinitionVersion=" + this.getProcessDefinitionVersion() + ", processDefinitionId="
+                + this.getProcessDefinitionId() + ", businessKey=" + this.getBusinessKey() + ", startTime="
+                + this.getStartTime() + ", endTime=" + this.getEndTime() + ", durationInMillis="
+                + this.getDurationInMillis() + ", startUserId=" + this.getStartUserId() + ", startActivityId="
+                + this.getStartActivityId() + ", deleteReason=" + this.getDeleteReason() + ", tenantId="
+                + this.getTenantId() + ", state=" + this.getState() + ")";
     }
 }
