@@ -1,16 +1,18 @@
 // Generated by delombok at Fri Jan 04 11:18:40 CET 2019
 package de.viadee.camunda.kafka.pollingclient.service.polling.rest.response;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import org.apache.commons.lang3.StringUtils;
+
 import java.util.Map;
 import java.util.Objects;
-import org.apache.commons.lang3.StringUtils;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 /**
  * Response structure of Camunda REST API <code>GET /history/variable-instance</code>
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class GetHistoricVariableInstancesResponse {
+
     private String id;
     private String name;
     private String type;
@@ -34,7 +36,8 @@ public class GetHistoricVariableInstancesResponse {
      * If {@link #valueInfo} is available as Map-Data, retrieve value of given key as String.
      */
     public String getValueInfoEntry(String key) {
-        return valueInfo instanceof Map ? StringUtils.trimToNull(Objects.toString(((Map) valueInfo).get(key), null)) : null;
+        return valueInfo instanceof Map ? StringUtils.trimToNull(Objects.toString(((Map) valueInfo).get(key), null))
+                : null;
     }
 
     @java.lang.SuppressWarnings("all")
@@ -220,6 +223,14 @@ public class GetHistoricVariableInstancesResponse {
     @java.lang.Override
     @java.lang.SuppressWarnings("all")
     public java.lang.String toString() {
-        return "GetHistoricVariableInstancesResponse(id=" + this.getId() + ", name=" + this.getName() + ", type=" + this.getType() + ", value=" + this.getValue() + ", valueInfo=" + this.getValueInfo() + ", processDefinitionKey=" + this.getProcessDefinitionKey() + ", processDefinitionId=" + this.getProcessDefinitionId() + ", processInstanceId=" + this.getProcessInstanceId() + ", executionId=" + this.getExecutionId() + ", activityInstanceId=" + this.getActivityInstanceId() + ", caseDefinitionKey=" + this.getCaseDefinitionKey() + ", caseDefinitionId=" + this.getCaseDefinitionId() + ", caseInstanceId=" + this.getCaseInstanceId() + ", caseExecutionId=" + this.getCaseExecutionId() + ", taskId=" + this.getTaskId() + ", tenantId=" + this.getTenantId() + ", errorMessage=" + this.getErrorMessage() + ", state=" + this.getState() + ")";
+        return "GetHistoricVariableInstancesResponse(id=" + this.getId() + ", name=" + this.getName() + ", type="
+                + this.getType() + ", value=" + this.getValue() + ", valueInfo=" + this.getValueInfo()
+                + ", processDefinitionKey=" + this.getProcessDefinitionKey() + ", processDefinitionId="
+                + this.getProcessDefinitionId() + ", processInstanceId=" + this.getProcessInstanceId()
+                + ", executionId=" + this.getExecutionId() + ", activityInstanceId=" + this.getActivityInstanceId()
+                + ", caseDefinitionKey=" + this.getCaseDefinitionKey() + ", caseDefinitionId="
+                + this.getCaseDefinitionId() + ", caseInstanceId=" + this.getCaseInstanceId() + ", caseExecutionId="
+                + this.getCaseExecutionId() + ", taskId=" + this.getTaskId() + ", tenantId=" + this.getTenantId()
+                + ", errorMessage=" + this.getErrorMessage() + ", state=" + this.getState() + ")";
     }
 }
